피드백 수정

+++ app/src/main/java/com/example/basic1/Text1Activity.kt
@@ -15,8 +15,6 @@ import kotlinx.coroutines.Job
 
 class Text1Activity : AppCompatActivity() {
     private var job: Job? = null
-
-    //    val randomValue = (1..100).random()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_test1)
@@ -25,7 +23,6 @@ class Text1Activity : AppCompatActivity() {
         setupButton()
         setJobAndLaunch()
     }
-
     private fun setupButton() {
         val button = findViewById<Button>(R.id.clickButton)
         button.setOnClickListener {
@@ -34,15 +31,11 @@ class Text1Activity : AppCompatActivity() {
             checkAnswerAndShowToast()
         }
     }
-
     private fun setRandomValueBetweenOneToHundred() {
         val randomTextView = findViewById<TextView>(R.id.textViewRandom)
         val randomValue = (1..100).random()
         randomTextView.setText(randomValue.toString())
-//      randomTextView.text = randomValue.toString()
-//      setText > EditText(입력받는 위젯)에 쓰이는 함수
     }
-
     private fun setJobAndLaunch() {
         val textView = findViewById<TextView>(R.id.spartaTextView)
 
@@ -50,25 +43,15 @@ class Text1Activity : AppCompatActivity() {
             for (i in 1..100) {
                 if (isActive) {
                     textView.text = i.toString()
-//                    if (i == randomValue) {
-//                        toast()
-//                        break
-//                    }
                     delay(500)
                 }
             }
-
         }
     }
-
-//    private fun toast() {
-//        Toast.makeText(this, "같은 값 입니다", Toast.LENGTH_SHORT).show()
-//    }
-
     private fun checkAnswerAndShowToast() {
         val textView = findViewById<TextView>(R.id.spartaTextView)
         val randomTextView = findViewById<TextView>(R.id.textViewRandom)
-        if (textView.text.toString().toInt() == randomTextView.text.toString().toInt()) {
+        if (textView.text.toString() == randomTextView.text.toString()) {
             Toast.makeText(this, "같은 값 입니다", Toast.LENGTH_SHORT).show()
         }
         Toast.makeText(this,"같은 값이 아닙니다",Toast.LENGTH_SHORT).show()
